requirements.txt
cbor
fido2
python-ldap